# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cinefiles Audio Copyright Clearance Service",
    "description": "Audio copyright clearance using AudD fingerprinting, Sync and Master fee estimates, audio clearance drafts, and plain-text script URL scanning.",
    "contact": {
      "name": "Cinefiles",
      "url": "https://github.com/Cinefiles"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/clearance/audio": {
      "post": {
        "tags": [
          "Clearance"
        ],
        "summary": "Request Audio Clearance",
        "operationId": "request_audio_clearance_api_v1_clearance_audio_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClearanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clearance/audio/direct": {
      "post": {
        "tags": [
          "Clearance"
        ],
        "summary": "Direct Audio Clearance",
        "operationId": "direct_audio_clearance_api_v1_clearance_audio_direct_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectClearanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clearance/audio/upload": {
      "post": {
        "tags": [
          "Clearance"
        ],
        "summary": "Upload Audio Clearance",
        "operationId": "upload_audio_clearance_api_v1_clearance_audio_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_audio_clearance_api_v1_clearance_audio_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scripts/scan": {
      "post": {
        "tags": [
          "Script Scanning"
        ],
        "summary": "Scan Script",
        "operationId": "scan_script_api_v1_scripts_scan_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_scan_script_api_v1_scripts_scan_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ui": {
      "get": {
        "summary": "Serve Ui",
        "operationId": "serve_ui_ui_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuddMatch": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "apple_music_link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apple Music Link"
          }
        },
        "type": "object",
        "required": [
          "title",
          "artist"
        ],
        "title": "AuddMatch"
      },
      "Body_scan_script_api_v1_scripts_scan_post": {
        "properties": {
          "script_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Script File"
          }
        },
        "type": "object",
        "required": [
          "script_file"
        ],
        "title": "Body_scan_script_api_v1_scripts_scan_post"
      },
      "Body_upload_audio_clearance_api_v1_clearance_audio_upload_post": {
        "properties": {
          "audio_file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Audio File"
          }
        },
        "type": "object",
        "required": [
          "audio_file"
        ],
        "title": "Body_upload_audio_clearance_api_v1_clearance_audio_upload_post"
      },
      "ClearanceRequest": {
        "properties": {
          "audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Url",
            "examples": [
              "https://example.com/clip.mp3"
            ]
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path",
            "examples": [
              "/uploads/clip.mp3"
            ]
          }
        },
        "type": "object",
        "title": "ClearanceRequest"
      },
      "ClearanceResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "pending",
              "denied"
            ],
            "title": "Status"
          },
          "royalty_free": {
            "type": "boolean",
            "title": "Royalty Free"
          },
          "match": {
            "$ref": "#/components/schemas/AuddMatch"
          },
          "licenses": {
            "items": {
              "$ref": "#/components/schemas/LicenseFee"
            },
            "type": "array",
            "title": "Licenses"
          },
          "total_fee_usd": {
            "type": "number",
            "title": "Total Fee Usd"
          },
          "currency": {
            "type": "string",
            "const": "USD",
            "title": "Currency"
          },
          "requested_at": {
            "type": "string",
            "title": "Requested At"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "contract_draft": {
            "$ref": "#/components/schemas/PdfDraftPayload"
          }
        },
        "type": "object",
        "required": [
          "status",
          "royalty_free",
          "match",
          "licenses",
          "total_fee_usd",
          "currency",
          "requested_at",
          "service",
          "version",
          "contract_draft"
        ],
        "title": "ClearanceResponse"
      },
      "DirectClearanceRequest": {
        "properties": {
          "song_title": {
            "type": "string",
            "minLength": 1,
            "title": "Song Title",
            "examples": [
              "Bohemian Rhapsody"
            ]
          },
          "artist": {
            "type": "string",
            "minLength": 1,
            "title": "Artist",
            "examples": [
              "Queen"
            ]
          },
          "timestamp_start": {
            "type": "string",
            "minLength": 1,
            "title": "Timestamp Start",
            "examples": [
              "00:01:30"
            ]
          },
          "timestamp_end": {
            "type": "string",
            "minLength": 1,
            "title": "Timestamp End",
            "examples": [
              "00:03:45"
            ]
          }
        },
        "type": "object",
        "required": [
          "song_title",
          "artist",
          "timestamp_start",
          "timestamp_end"
        ],
        "title": "DirectClearanceRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LicenseFee": {
        "properties": {
          "license_type": {
            "type": "string",
            "enum": [
              "Sync",
              "Master"
            ],
            "title": "License Type"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "amount_usd": {
            "type": "number",
            "title": "Amount Usd"
          }
        },
        "type": "object",
        "required": [
          "license_type",
          "description",
          "amount_usd"
        ],
        "title": "LicenseFee"
      },
      "PdfDraftPayload": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "const": "application/pdf",
            "title": "Content Type"
          },
          "data_base64": {
            "type": "string",
            "title": "Data Base64"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "content_type",
          "data_base64",
          "page_count"
        ],
        "title": "PdfDraftPayload"
      },
      "ScriptScanResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "candidate_count": {
            "type": "integer",
            "title": "Candidate Count"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ScriptUrlCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "candidate_count",
          "candidates",
          "message"
        ],
        "title": "ScriptScanResponse"
      },
      "ScriptUrlCandidate": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "line_number": {
            "type": "integer",
            "title": "Line Number"
          },
          "context": {
            "type": "string",
            "title": "Context"
          },
          "likely_audio": {
            "type": "boolean",
            "title": "Likely Audio"
          },
          "media_hint": {
            "type": "string",
            "enum": [
              "audio",
              "video",
              "unknown"
            ],
            "title": "Media Hint"
          }
        },
        "type": "object",
        "required": [
          "url",
          "line_number",
          "context",
          "likely_audio",
          "media_hint"
        ],
        "title": "ScriptUrlCandidate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}